Set multiuser installation context to all users
Fixes #3
This explicitly requires administrator level rights on the installer and notifies the user that install may fail if rights are not present.
The installer is already run with admin rights on Vista+ (Windows detects NSIS as an installer and, lacking any metadata, assumes admin is required).
Furthermore, installation to Program Files and HKLM fail silently without admin rights.